+++ com.unity.render-pipelines.universal/Editor/GlobalSettings/UniversalRenderPipelineGlobalSettingsUI.Skin.cs
@@ -19,6 +19,7 @@ namespace UnityEditor.Rendering.Universal
             public static readonly GUIContent stripDebugVariantsLabel = EditorGUIUtility.TrTextContent("Strip Debug Variants", "When disabled, all debug display shader variants are removed when you build for the Unity Player. This decreases build time, but prevents the use of Rendering Debugger in Player builds.");
             public static readonly GUIContent stripUnusedPostProcessingVariantsLabel = EditorGUIUtility.TrTextContent("Strip Unused Post Processing Variants", "Controls whether strips automatically post processing shader variants based on VolumeProfile components. It strips based on VolumeProfiles in project and not scenes that actually uses it.");
             public static readonly GUIContent stripUnusedVariantsLabel = EditorGUIUtility.TrTextContent("Strip Unused Variants", "Controls whether strip disabled keyword variants if the feature is enabled.");
+            public static readonly GUIContent stripUnusedLODCrossFadeVariantsLabel = EditorGUIUtility.TrTextContent("Strip Unused LOD Cross Fade Variants", "Controls whether strip off variants if the LOD Cross Fade feature is disabled. It strips based on the UniversalRenderingPipelineAsset.enableLODCrossFade property.");
         }
     }
 }