+++ com.unity.render-pipelines.universal/Runtime/UniversalRenderPipelineGlobalSettings.cs
@@ -322,6 +322,8 @@ namespace UnityEngine.Rendering.Universal
 
         [SerializeField] bool m_StripUnusedVariants = true;
 
+        [SerializeField] bool m_StripUnusedLODCrossFadeVariants = true;
+
         /// <summary>
         /// Controls whether debug display shaders for Rendering Debugger are available in Player builds.
         /// </summary>
@@ -343,6 +345,12 @@ namespace UnityEngine.Rendering.Universal
         /// Controls whether strip off variants if the feature is enabled.
         /// </summary>
         public bool stripUnusedVariants { get => m_StripUnusedVariants; set { m_StripUnusedVariants = value; } }
+
+        /// <summary>
+        /// Controls whether strip off variants if the LOD Cross Fade feature is disabled.
+        /// It strips based on UniversalRenderingPipelineAsset.enableLODCrossFade property.
+        /// </summary>
+        public bool stripUnusedLODCrossFadeVariants { get => m_StripUnusedLODCrossFadeVariants; set { m_StripUnusedLODCrossFadeVariants = value; } }
 
         #endregion
     }