+++ com.unity.render-pipelines.universal/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- All builtin URP shaders as long as URP ShaderGraph shaders will support Mesh LOD CrossFade which type can be selected in UniversalRenderPipelineAsset.lodCrossFadeDitheringType property.
+
 ## [14.0.2] - 2021-02-04
 
 ### Added